StoryTime-Generator/requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0
python-dotenv>=1.0.0
langchain-mcp-adapters>=0.1.0
huggingface_hub>=0.20.0
Pillow>=10.0.0
